requirements.txt
azure-storage-file-datalake==12.12.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
pytz